requirements.txt
bcrypt==3.2.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
Flask-Cors==5.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pycparser==2.22
PyJWT==2.1.0
pymongo==3.11.3
python-dotenv==1.1.0
python-http-client==3.3.7
requests==2.32.3
sendgrid==6.11.0
six==1.17.0
starkbank-ecdsa==2.2.0
urllib3==2.3.0
Werkzeug==3.1.3
